Fix arrays of object fields
Yet another instance of issue #375.

Specifically, any array that contained non-immediate descendants (e.g.
arrayField[].deepObjectField, or arrayField[][], but not plain old
arrayField[]) would remain the same instance if any of its descendant
fields were updated but no items were added to or removed from the
top-level array.

The issue was that we were only making a fresh copy of the array (that's
`changed = true`) in the case where it a raw-value array item was
changed, but not in the general case.